pyproject.toml
[project]
name = "bionote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "groq>=0.28.0",
    "load-dotenv>=0.1.0",
    "pillow>=11.2.1",
    "selenium>=4.33.0",
]
